+++ src/sink/checks.jl
@@ -11,11 +11,13 @@ additional checks on the data.
 - The field `cap` is required to be non-negative.
 - The values of the dictionary `input` are required to be non-negative.
 - The dictionary `penalty` is required to have the keys `:deficit` and `:surplus`.
+- The values `:deficit` and `:surplus` of the dictionary `penalty` are required to be
+  indexable by a `PeriodPartition`.
 - The sum of the values `:deficit` and `:surplus` in the dictionary `penalty` has to be
   non-negative to avoid an infeasible model.
-- The remainder of the divison of the lowest time structure by the period length must be 0.
-- The length of the period demand must equal the length of the lowest period times the
-  parameter period length.
+- The individual periods must all satisfy the specified duration(s).
+- The field `period_demand` is required to be non-negative and indexable by a
+  `PeriodPartition`.
 """
 function EMB.check_node(
     n::PeriodDemandSink,
@@ -23,52 +25,59 @@ function EMB.check_node(
     modeltype::EnergyModel,
     check_timeprofiles::Bool,
 )
-    EMB.check_node_default(n, 𝒯, modeltype, check_timeprofiles)
+    𝒯ᵖᵈ = periods(n, 𝒯)
+    per_dur = period_duration(n)
+    bool = true
 
-    # Check that the period length and demand is working with the time structure
-    # The check will only be activated in 0.3
-    # for (idx_sp, ts_oper) ∈ enumerate(𝒯.operational)
-    #     sub_msg = "the operational time structure in strategic period $(idx_sp)"
-    #     check_period_ts(ts_oper, n, sub_msg)
-    # end
-end
-
-"""
-    check_period_ts(ts::RepresentativePeriods, n::PeriodDemandSink, msg::String)
-    check_period_ts(ts::OperationalScenarios, n::PeriodDemandSink, msg::String)
-    check_period_ts(ts::SimpleTimes, n::PeriodDemandSink, msg::String)
-
-Function for checking that the timestructure is valid in combination with the chosen period
-structure in a [`PeriodDemandSink`(@ref).
-"""
-function check_period_ts(ts::RepresentativePeriods, n::PeriodDemandSink, msg::String)
-    for (idx, ts_oper) ∈ enumerate(ts.rep_periods)
-        sub_msg = msg * " in representative period $(idx)"
-        check_period_ts(ts_oper, n, sub_msg)
-    end
-end
-function check_period_ts(ts::OperationalScenarios, n::PeriodDemandSink, msg::String)
-    for (idx, ts_oper) ∈ enumerate(ts.scenarios)
-        sub_msg = msg * " in operational scenario $(idx)"
-        check_period_ts(ts_oper, n, sub_msg)
-    end
-end
-function check_period_ts(ts::SimpleTimes, n::PeriodDemandSink, msg::String)
-    len = period_length(n)
-    n_per = number_of_periods(n)
     @assert_or_log(
-        length(ts)%len == 0,
-        "The specified period length does not work with $(msg)."
+        all(capacity(n, t) ≥ 0 for t ∈ 𝒯),
+        "The capacity must be non-negative."
     )
     @assert_or_log(
-        length(ts)/len ≤ n_per,
-        "The vector `period_demand` is shorter than the $(msg)."
+        all(inputs(n, p) ≥ 0 for p ∈ inputs(n)),
+        "The values for the Dictionary `input` must be non-negative."
     )
-    if length(ts)%len == 0 & length(ts)/len < n_per
-        @warn(
-            "The vector `period_demand` is longer than required in $(msg). " *
-            "The last $(Int(n_per-length(ts)/len)) values will be omitted.",
-            maxlog=1,
+    @assert_or_log(
+        :surplus ∈ keys(n.penalty) && :deficit ∈ keys(n.penalty),
+        "The entries `:surplus` and `:deficit` are required in the dictionary `penalty`."
+    )
+    if :surplus ∈ keys(n.penalty)
+        message = "are not allowed for the key `:surplus` in the dictionary `penalty`."
+        bool *= EMB.check_partition_profile(surplus_penalty(n), message)
+    else
+        bool = false
+    end
+    if :deficit ∈ keys(n.penalty)
+        message = "are not allowed for the key `:deficit` in the dictionary `penalty`."
+        bool *= EMB.check_partition_profile(deficit_penalty(n), message)
+    else
+        bool = false
+    end
+
+    if bool
+        @assert_or_log(
+            all(surplus_penalty(n, t_pd) + deficit_penalty(n, t_pd) ≥ 0 for t_pd ∈ 𝒯ᵖᵈ),
+            "An inconsistent combination of `:surplus` and `:deficit` leads to an infeasible model."
+        )
+    end
+
+    message = "are not allowed for the field `:period_duration`."
+    bool = EMB.check_partition_profile(period_duration(n), message)
+    if bool
+        @assert_or_log(
+            all(sum(duration(t) for t ∈ t_pd) ≥ per_dur[t_pd] for t_pd ∈ 𝒯ᵖᵈ),
+            "The duration of the last period on the `SimpleTimes` level is shorter than " *
+            "specified. This is caused by inconsistently specified `period_duration` and" *
+            "time structure."
+        )
+    end
+
+    message = "are not allowed for the field `:period_demand`."
+    bool = EMB.check_partition_profile(period_demand(n), message)
+    if bool
+        @assert_or_log(
+            all(period_demand(n, t_pd) ≥ 0 for t_pd ∈ 𝒯ᵖᵈ),
+            "The period demand must be non-negative."
         )
     end
 end